# docs/swagger.yaml
basePath: /
definitions:
  task.CreateTaskRequest:
    properties:
      desc:
        type: string
      label:
        type: string
      title:
        type: string
      userId:
        type: string
    required:
    - label
    - title
    - userId
    type: object
  task.EditTaskRequest:
    properties:
      desc:
        type: string
      label:
        type: string
      title:
        type: string
    required:
    - label
    - title
    type: object
  task.Task:
    properties:
      createdAt:
        type: string
      desc:
        type: string
      id:
        type: string
      label:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  user.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  user.RegisterUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  user.UserResponse:
    properties:
      apiKey:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Taskly backend API
  title: Taskly API
  version: "1.0"
paths:
  /task:
    get:
      description: Get all tasks
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Limit per page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/task.Task'
            type: array
      summary: Get tasks
      tags:
      - task
    post:
      description: Create a new task
      parameters:
      - description: Create Task Request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/task.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
      summary: Create task
      tags:
      - task
  /task/{id}:
    delete:
      description: Delete task by id
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
      summary: Delete task
      tags:
      - task
    get:
      description: Get task by id
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
      summary: Get task
      tags:
      - task
    put:
      description: Edit task by id
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Edit Task Request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/task.EditTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
      summary: Edit task
      tags:
      - task
  /user/login:
    post:
      description: Login a user
      parameters:
      - description: Login User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
      summary: Login user
      tags:
      - user
  /user/register:
    post:
      description: Register a new user
      parameters:
      - description: Register User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
      summary: Register user
      tags:
      - user
swagger: "2.0"